requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
idna==3.4
paho-mqtt==1.6.1
pycryptodome==3.18.0
requests==2.31.0
tinytuya==1.12.9
urllib3==2.0.3
pymongo==4.6.0